+++ test_csv_permissions/test_legacy.py
@@ -7,11 +7,11 @@ from django.test import TestCase
 import csv_permissions
 import csv_permissions.legacy
 import csv_permissions.permissions
+from csv_permissions.test_utils import override_csv_permissions
 import test_csv_permissions.rules
 
 from .models import TestModelA
 from .models import TestModelE
-from .test_utils import override_csv_permissions
 from .test_utils import USER1_TYPE
 from .test_utils import User1Factory
 from .test_utils import USER2_TYPE
