+++ test_csv_permissions/test_parse.py
@@ -7,9 +7,9 @@ from django.test import TestCase
 
 import csv_permissions.evaluators
 import csv_permissions.permissions
+from csv_permissions.test_utils import override_csv_permissions
 
 from .models import TestModelA
-from .test_utils import override_csv_permissions
 from .test_utils import USER1_TYPE
 from .test_utils import User1Factory
 from .test_utils import USER2_TYPE
